chatbot/kenokim/frontend/requirements.txt
streamlit
requests
python-dotenv
nest-asyncio
asyncio 